// apps/web/src/shared/api/generated/model/unitProgressSummaryResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */
import type { UnitProgressSummaryResponseStatus } from './unitProgressSummaryResponseStatus';

export interface UnitProgressSummaryResponse {
  unitId: number;
  title: string;
  status: UnitProgressSummaryResponseStatus;
}