// apps/web/src/shared/api/generated/model/unitProgressSummaryResponseStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */

export type UnitProgressSummaryResponseStatus =
  (typeof UnitProgressSummaryResponseStatus)[keyof typeof UnitProgressSummaryResponseStatus];

export const UnitProgressSummaryResponseStatus = {
  NOT_STARTED: 'NOT_STARTED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
} as const;